@@ -18,13 +18,18 @@
         "Memio\\TwigTemplateEngine\\Config\\": "config"
     }},
     "require": {
-        "memio/model": "^1.3",
-        "memio/pretty-printer": "^1.0",
-        "php": "^5.3 || ^7.0",
+        "memio/model": "^2.0",
+        "memio/pretty-printer": "^2.0",
+        "php": "^7.0",
         "twig/twig": "^1.18 || ^2.0"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^1.6",
-        "phpspec/phpspec": "^2.4 || ^3.0"
+        "friendsofphp/php-cs-fixer": "^2.10",
+        "phpspec/phpspec": "^4.3"
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "2.1-dev"
+        }
     }
 }
