@@ -11,7 +11,8 @@
 
 namespace spec\Memio\TwigTemplateEngine\TwigExtension\Line;
 
-use Memio\Model\Object;
+use Memio\Model\Objekt;
+use Memio\TwigTemplateEngine\TwigExtension\Line\LineStrategy;
 use PhpSpec\ObjectBehavior;
 
 class ObjectLineStrategySpec extends ObjectBehavior
@@ -21,37 +22,39 @@ class ObjectLineStrategySpec extends ObjectBehavior
 
     function it_is_a_line_strategy()
     {
-        $this->shouldImplement('Memio\TwigTemplateEngine\TwigExtension\Line\LineStrategy');
+        $this->shouldImplement(LineStrategy::class);
     }
 
-    function it_supports_objects(Object $object)
+    function it_supports_objects(Objekt $object)
     {
         $this->supports($object)->shouldBe(true);
     }
 
-    function it_needs_line_after_constants_if_object_has_both_constants_and_properties(Object $object)
-    {
-        $object->allConstants()->willReturn(array(1));
-        $object->allProperties()->willReturn(array(2));
-        $object->allMethods()->willReturn(array());
+    function it_needs_line_after_constants_if_object_has_both_constants_and_properties(
+        Objekt $object
+    ) {
+        $object->allConstants()->willReturn([1]);
+        $object->allProperties()->willReturn([2]);
+        $object->allMethods()->willReturn([]);
 
         $this->needsLineAfter($object, self::CONSTANT_BLOCK)->shouldBe(true);
     }
 
-    function it_needs_line_after_constants_if_object_has_both_constants_and_methods(Object $object)
-    {
-        $object->allConstants()->willReturn(array(1));
-        $object->allProperties()->willReturn(array());
-        $object->allMethods()->willReturn(array(2));
+    function it_needs_line_after_constants_if_object_has_both_constants_and_methods(
+        Objekt $object
+    ) {
+        $object->allConstants()->willReturn([1]);
+        $object->allProperties()->willReturn([]);
+        $object->allMethods()->willReturn([2]);
 
         $this->needsLineAfter($object, self::CONSTANT_BLOCK)->shouldBe(true);
     }
 
-    function it_needs_line_after_properties_if_object_has_both_properties_and_methods(Object $object)
+    function it_needs_line_after_properties_if_object_has_both_properties_and_methods(Objekt $object)
     {
-        $object->allConstants()->willReturn(array());
-        $object->allProperties()->willReturn(array(1));
-        $object->allMethods()->willReturn(array(2));
+        $object->allConstants()->willReturn([]);
+        $object->allProperties()->willReturn([1]);
+        $object->allMethods()->willReturn([2]);
 
         $this->needsLineAfter($object, self::PROPERTY_BLOCK)->shouldBe(true);
     }
