@@ -16,22 +16,16 @@ use Memio\Model\File;
 
 class FileLineStrategy implements LineStrategy
 {
-    /**
-     * {@inheritDoc}
-     */
-    public function supports($model)
+    public function supports($model): bool
     {
         return $model instanceof File;
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function needsLineAfter($model, $block)
+    public function needsLineAfter($model, string $block): bool
     {
         $fullyQualifiedNames = $model->allFullyQualifiedNames();
         if ('fully_qualified_names' === $block) {
-            return (!empty($fullyQualifiedNames));
+            return !empty($fullyQualifiedNames);
         }
 
         throw new InvalidArgumentException('The function needs_line_after does not support given "'.$block.'"');
